Explicit casting and logging INFO at Production

+++ grails-app/controllers/org/etive/city4age/repository/ProximityEventController.groovy
@@ -19,7 +19,7 @@ class ProximityEventController {
             if (beacon && device && device.careReceiver == careReceiver) {
                 def event = proximityEventService.createProximityEvent(careReceiver, beacon, device, json)
                 // Note that we have heard from the device
-                deviceService.updateLastContact(device, json.timestamp)
+                deviceService.updateLastContact(device, json.timestamp as Long)
                 respond(event, status: 201)
             }
             else {
